Refactor cached fibonacci
It uses a *callback* approach. Before, it was the Fibonacci library who was responsible of checking if the value was in the cache or not.

Now, it passes a callback to the cache. If the value is not present, the cache runs the callback and caches the computed value. If it's present, just returns the value.